gcc on Mac OSX complains conflicting types for ncmpiget_leng
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1585 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/utils/ncmpigen/genlib.h
@@ -83,7 +83,7 @@ extern int ncmpiwrap(void);
 extern int ncmpiget_lineno(void);
 extern FILE *ncmpiget_in(void);
 extern FILE *ncmpiget_out(void);
-extern int ncmpiget_leng(void);
+extern size_t ncmpiget_leng(void);
 extern char *ncmpiget_text(void);
 extern void ncmpiset_lineno(int  line_number);
 extern void ncmpiset_in(FILE *in_str);
